finished auth and sessions

+++ motherducker/connections/views.py
@@ -14,6 +14,12 @@ class HomePageView(TemplateView):
     template_name = 'index.html'
 
     def get(self, request, *args, **kwargs):
+        print(request.session)
+        try:
+            if request.session['session']['username'] and request.session['session']['password']:
+                request.session.flush()
+        except KeyError:
+            pass
         context = {}
         return render(request, self.template_name, context)
 
@@ -90,19 +96,28 @@ class TerminalView(TemplateView):
 
     def get(self, request, *args, **kwargs):
         time.sleep(3)
-        context = {}
-        context['terminal'] = 'C:\\>'
-        context['uuid'] = self.kwargs.get('uuid')
-        context['history'] = TerminalHistory.objects.filter(connection=self.kwargs.get('uuid'))
-        get_uuid = Connection.objects.get(uuid=self.kwargs.get('uuid'))
         try:
-            terminal_data = TerminalLog.objects.filter(connection=get_uuid).latest('id')
-            context['terminal'] = terminal_data.current_directory
-            context['terminal_output'] = terminal_data.content
-            context['terminal_history'] = TerminalLog.objects.filter(connection=get_uuid)
-        except:
-            pass
-        return render(request, self.template_name, context)
+            username = request.session['session']['username']
+            password = request.session['session']['password']
+            user = authenticate(request, username=username, password=password)
+            if user is not None:
+                context = {'username': request.session['session']['username']}
+                context['terminal'] = 'C:\\>'
+                context['uuid'] = self.kwargs.get('uuid')
+                context['history'] = TerminalHistory.objects.filter(connection=self.kwargs.get('uuid'))
+                get_uuid = Connection.objects.get(uuid=self.kwargs.get('uuid'))
+                try:
+                    terminal_data = TerminalLog.objects.filter(connection=get_uuid).latest('id')
+                    context['terminal'] = terminal_data.current_directory
+                    context['terminal_output'] = terminal_data.content
+                    context['terminal_history'] = TerminalLog.objects.filter(connection=get_uuid)
+                except:
+                    pass
+                return render(request, self.template_name, context)
+            else:
+                return redirect('home')
+        except KeyError:
+            return redirect('home')
 
     def post(self, request, *args, **kwargs):
         context = {}
@@ -138,39 +153,82 @@ class TerminalView(TemplateView):
         return TerminalView.get(self, request, *args, **kwargs)
 
 
-
 class ScriptsView(TemplateView):
     template_name = 'scripts.html'
 
     def get(self, request, *args, **kwargs):
-        context = super(ScriptsView, self).get_context_data(**kwargs)
-        context['payloads'] = Payload.objects.all()
-        return render(request, self.template_name, context)
+
+        try:
+            username = request.session['session']['username']
+            password = request.session['session']['password']
+            user = authenticate(request, username=username, password=password)
+            if user is not None:
+                context = super(ScriptsView, self).get_context_data(**kwargs)
+                context['payloads'] = Payload.objects.all()
+                context['username'] = request.session['session']['username']
+                return render(request, self.template_name, context)
+            else:
+                return redirect('home')
+        except KeyError:
+            return redirect('home')
+
 
     def post(self, request, *args, **kwargs):
-        context = {'payloads': Payload.objects.all(), 'uuid': self.kwargs.get('uuid')}
-        get_uuid = Connection.objects.get(uuid=self.kwargs.get('uuid'))
-        ScriptData.objects.all().delete()
-        ScriptData.objects.create(input=request.POST.get('payload'),
-                                  payload_name=request.POST.get('payload_name'),
-                                  connection_id=get_uuid)
-        return render(request, self.template_name, context)
+        try:
+            username = request.session['session']['username']
+            password = request.session['session']['password']
+            user = authenticate(request, username=username, password=password)
+            if user is not None:
+                context = {'username': request.session['session']['username']}
+                context['payloads'] = Payload.objects.all()
+                context['uuid'] = self.kwargs.get('uuid')
+                get_uuid = Connection.objects.get(uuid=self.kwargs.get('uuid'))
+                ScriptData.objects.all().delete()
+                ScriptData.objects.create(input=request.POST.get('payload'),
+                                          payload_name=request.POST.get('payload_name'),
+                                          connection_id=get_uuid)
+                return render(request, self.template_name, context)
+            else:
+                return redirect('home')
+        except KeyError:
+            return redirect('home')
 
 
 class ConnectionDetailsView(TemplateView):
     template_name = 'connection_details.html'
 
     def get(self, request, *args, **kwargs):
-        context = super(ConnectionDetailsView, self).get_context_data(**kwargs)
-        context['uuid'] = self.kwargs.get('uuid')
-        context['connection'] = Connection.objects.get(uuid=context['uuid'])
-        context['terminal_log'] = TerminalLog.objects.filter(connection=context['connection'])
-        context['script_log'] = ScriptLog.objects.filter(connection=context['connection'])
-        return render(request, self.template_name, context)
+        try:
+            username = request.session['session']['username']
+            password = request.session['session']['password']
+            user = authenticate(request, username=username, password=password)
+            if user is not None:
+                context = super(ConnectionDetailsView, self).get_context_data(**kwargs)
+                context['username'] = request.session['session']['username']
+                context['uuid'] = self.kwargs.get('uuid')
+                context['connection'] = Connection.objects.get(uuid=context['uuid'])
+                context['terminal_log'] = TerminalLog.objects.filter(connection=context['connection'])
+                context['script_log'] = ScriptLog.objects.filter(connection=context['connection'])
+                return render(request, self.template_name, context)
+            else:
+                return redirect('home')
+        except KeyError:
+            return redirect('home')
 
     def post(self, request, *args, **kwargs):
-        context = {'uuid': self.kwargs.get('uuid')}
-        return render(request, self.template_name, context)
+
+        try:
+            username = request.session['session']['username']
+            password = request.session['session']['password']
+            user = authenticate(request, username=username, password=password)
+            if user is not None:
+                context = {'username': request.session['session']['username']}
+                context['uuid'] = self.kwargs.get('uuid')
+                return render(request, self.template_name, context)
+            else:
+                return redirect('home')
+        except KeyError:
+            return redirect('home')
 
 
 class RegisterView(TemplateView):
